feat: export symbol-id as a submodule
The `symbolIdentifier` function is a base jsii feature that other tools can depend on without having to depend on the rest of jsii.
Currently a dependant package has to import all of the jsii code just to get `symbolIdentifier`.
This can have performance implications, as well as producing larger bundles.

This change exports `symbolIdentifier` as a submodule to allow a direct import.
