+++ package.json
@@ -93,7 +93,8 @@
   "exports": {
     ".": "./lib/index.js",
     "./bin/jsii": "./bin/jsii",
-    "./package.json": "./package.json"
+    "./package.json": "./package.json",
+    "./common": "./lib/common/index.js"
   },
   "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\"."
 }
