+++ src/common/index.ts
@@ -1,0 +1,1 @@
+export { symbolIdentifier } from './symbol-id';
