+++ src/common/symbol-id.ts
@@ -1,9 +1,9 @@
 import * as fs from 'node:fs';
 import * as path from 'node:path';
-import { Assembly } from '@jsii/spec';
+import type { Assembly } from '@jsii/spec';
 import * as ts from 'typescript';
 
-import { findUp } from './utils';
+import { findUp } from './find-utils';
 
 /**
  * Additional options that may be provided to the symbolIdentifier.
